Rename package weather_model_graphs -> poseidon_graphs
Rename the importable module (src/weather_model_graphs -> src/poseidon_graphs)
and the distribution (weather-model-graphs -> poseidon-graphs), updating
pyproject name/packages, the importlib.metadata version lookup, all test
imports, and docs prose. Upstream attribution URLs (github.com/mllam/...,
mllam.github.io/...) and the historical CHANGELOG are left unchanged.
