+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "weather-model-graphs"
+name = "poseidon-graphs"
 version = "0.3.0"
 description = "Tools for creating and visualising data-driven weather model graphs"
 authors = [
@@ -7,7 +7,7 @@ authors = [
 ]
 
 [tool.setuptools]
-packages = ["weather_model_graphs"]
+packages = ["poseidon_graphs"]
 
 dependencies = [
     "numpy>=1.26.4",
