+++ src/poseidon_graphs/__init__.py
@@ -1,7 +1,7 @@
 import importlib.metadata
 
 try:
-    __version__ = importlib.metadata.version("weather-model-graphs")
+    __version__ = importlib.metadata.version("poseidon-graphs")
 except importlib.metadata.PackageNotFoundError:
     __version__ = "unknown"
