+++ src/poseidon_graphs/create/mesh/kinds/hierarchical.py
